Use permit for fetch
